+++ CHANGELOG.md
@@ -11,6 +11,16 @@ All user visible changes to this project will be documented in this file. This p
 
 [Diff](/../../compare/v0.1.0-alpha.8...v0.1.0-alpha.9) | [Milestone](/../../milestone/5)
 
+### Added
+
+- UI:
+    - Chat page:
+        - History clearing. ([#361])
+    - Chats tab:
+        - Multiple chats selection. ([#361], [#348])
+    - Contacts tab:
+        - Multiple contacts selection. ([#361], [#348])
+
 ### Changed
 
 - UI:
@@ -27,7 +37,9 @@ All user visible changes to this project will be documented in this file. This p
 [#258]: /../../issues/258
 [#264]: /../../issues/264
 [#270]: /../../pull/270
+[#348]: /../../issues/348
 [#357]: /../../pull/357
+[#361]: /../../pull/361
 [#362]: /../../pull/362
 [#356]: /../../pull/356